test: update cipher tests
